refactor: move participation code to participations folder
