+++ lab/models/__init__.py
@@ -1,4 +1,4 @@
 from ..objects import Location, Object, ObjectGroup, Period  # noqa: F401
+from ..participations.models import Institution, Participation  # noqa: F401
 from ..projects.models import BeamTimeRequest, Project  # noqa: F401
 from ..runs.models import Run  # noqa: F401
-from .participation import Institution, Participation  # noqa: F401
