+++ lab/participations/tests/test_institution_admiin.py
@@ -2,7 +2,7 @@ from django.contrib.auth import get_user_model
 from django.test import Client, TestCase
 from django.urls import reverse
 
-from ...models import Institution
+from ..models import Institution
 
 
 class TestInstitutionAdminViewAsParticipant(TestCase):
