+++ lab/projects/models.py
@@ -9,7 +9,7 @@ from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
 from slugify import slugify
 
-from lab.models.participation import Participation
+from lab.participations.models import Participation
 from lab.runs.models import Run
 from lab.validators import valid_filename
 from shared.models import TimestampedModel
