+++ lab/tests/widgets/test_auto_complete_widget.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ...models.participation import Institution
+from ...participations.models import Institution
 from ...widgets import AutoCompleteWidget
